IOS-6930: Add ManageTokensAdapter
